<Feat> <Fix> 1. alarms 앱 생성 및 각 앱에 있던 알림기능을 합치고 중복되는 컨슈머 삭제 2. alarms 앱 생성으로 인해 각 앱에 있던 알림관련 APIView 또는    시리얼라이저 이동 3. 알림 앱에 주석처리 한 부분은 추후에 다 같이 의논하면 좋을 것 같아서    기존 로직으로 커밋하고 주석처리 했습니당
 #214
