+++ shop/apps.py
@@ -4,6 +4,3 @@ from django.apps import AppConfig
 class ShopConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'shop'
-
-    def ready(self):
-        import shop.signals
