+++ users/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
-from .models import User, UserProfile, Notification
+from .models import User, UserProfile
 
 
 admin.site.register(User)
 admin.site.register(UserProfile)
-admin.site.register(Notification)
