+++ users/urls.py
@@ -32,8 +32,4 @@ urlpatterns = [
 
     # 유저프로필
     path('profile/', views.UserProfileAPIView.as_view(), name='user_profile'),
-
-    # 유저 알림 조회
-    path('notifications/', views.NotificationListAPIView.as_view(),
-         name='notification_list'),
 ]
