+++ users/views.py
@@ -1,4 +1,4 @@
-from users.models import User, UserProfile, Notification
+from users.models import User, UserProfile
 import requests
 import os
 import base64
@@ -20,7 +20,6 @@ from users.serializers import (
     ResetPasswordSerializer,
     ResetPasswordEmailSerializer,
     UserProfileSerializer,
-    UserNotificationSerializer,
     UserWithdrawalSerializer
 )
 from allauth.socialaccount.models import SocialAccount
@@ -508,47 +507,3 @@ class UserProfileAPIView(APIView):
         return Response(
             serializer.errors, status=status.HTTP_400_BAD_REQUEST
         )
-
-
-class NotificationListAPIView(APIView):
-    '''
-    작성자 : 장소은
-    내용 : 유저 캠페인 알림/재입고 알림 조회 기능, 개별 삭제/일괄 삭제 기능 
-    작성일 : 2023.06.22
-    '''
-    permission_classes = [IsAuthenticated]
-    pagination_class = CustomPagination
-
-    def get(self, request):
-        notifications = Notification.objects.filter(
-            user=request.user.id).order_by('-created_at')
-        print(notifications)
-        paginator = self.pagination_class()
-        result_page = paginator.paginate_queryset(notifications, request)
-        serializer = UserNotificationSerializer(
-            result_page,
-            many=True
-        )
-        return paginator.get_paginated_response(serializer.data)
-
-    def delete(self, request):
-        user = request.user.id
-        print(user)
-        notification_id = request.data.get('notification_id')
-        if notification_id:
-            try:
-                notification = Notification.objects.get(
-                    id=notification_id,
-                    user=user
-                )
-                print(notification)
-                notification.delete()
-                return Response(status=204)
-
-            except Notification.DoesNotExist:
-                return Response({'error': '알림내역을 찾을 수 없습니다.'}, status=404)
-
-        else:
-            notifications = Notification.objects.filter(user=user)
-            notifications.delete()
-            return Response(status=204)
